metrics-collector/requirements.txt
CherryPy==3.6.0
plan==0.5
crontab==0.20.2
crab==0.4.2
Mako>1.0
numpy
pandas
boto>=2.32
requests==2.2.1
python-dateutil==2.2
six==1.5.2
google-api-python-client==1.2
httplib2==0.8
python-gflags==2.0
pytz==2014.4
urllib3==1.7.1
wsgiref==0.1.2
elasticsearch==1.2.0
futures==2.2.0
SQLAlchemy==0.9.4
